requirements.txt
webdriver-manager==3.8.5
allure-pytest==2.13.1
python-dotenv==1.0.0
pytest-xdist==3.2.1
pytest-html==3.2.0
pytest-bdd==6.1.1
selenium==4.8.3
flake8==6.0.0
black==23.3.0
pytest==7.2.2
tox==4.4.11
